Calculate functional or pathway enrichment from a gradient

enrich_gradient() calculates over-representation statistics (ORA) using a ranking of genes from distance_based_analysis() For ORA using an external or already defined interest list of genes and gene sets, use leapR functions directly

enrich_gradient(
  spe,
  assay_name,
  geneset,
  feature_column,
  ranking_column,
  method = "ztest"
)

Arguments

spe

SpatialExperiment object containing spatial omics data and spatial diffex results

assay_name

name of assay

geneset

in GMT format

feature_column

Column of rowData that maps to gene set

ranking_column

Column of rowData that maps to ranks

method

Rank enrichmemnt method used by leapR. Default is ztest

Value

A dataframe containing results from over-representation analysis of members of gene sets in the interest list of genes based on filtering criteria above.

Examples

data(smallPancData)
data(pancMeta)
data(protMeta)
img0.spe <- convert_to_spe(smallPancData$Image_0,
  pancMeta,
  protMeta,
  feature_meta_colname = "pancProts",
  image_files = system.file("extdata", "Image_0.png", package = "spammR"),
  image_sample_ids = "Image0",
  spatial_coords_colnames = c("x_pixels", "y_pixels"),
  sample_id = "Image0",
  image_ids = "Image0"
)
#> Note: Only mapping metadata for 2986 features out of 3000 data points
img0.spe <- distance_based_analysis(img0.spe,
  "proteomics",
  sampleCategoryCol = "IsletOrNot",
  sampleCategoryValue = "Islet"
)
library(leapR)
data("krbpaths")
rank.res <- enrich_gradient(img0.spe,
  geneset = krbpaths,
  feature_column = "PrimaryGeneName",
  ranking_column = "IsletDistancespearmanCor"
)
head(rank.res)
#>                                                                                   ingroup_n
#> REACTOME_IMMUNOREGULATORY_INTERACTIONS_BETWEEN_A_LYMPHOID_AND_A_NON_LYMPHOID_CELL        16
#> REACTOME_RNA_POLYMERASE_I_PROMOTER_CLEARANCE                                              8
#> REACTOME_G_ALPHA_I_SIGNALLING_EVENTS                                                     15
#> REACTOME_STEROID_METABOLISM                                                              18
#> REACTOME_G_ALPHA_Q_SIGNALLING_EVENTS                                                     18
#> REACTOME_NRAGE_SIGNALS_DEATH_THROUGH_JNK                                                  9
#>                                                                                                                                                                                               ingroupnames
#> REACTOME_IMMUNOREGULATORY_INTERACTIONS_BETWEEN_A_LYMPHOID_AND_A_NON_LYMPHOID_CELL                   C3, IGKC, HLA-B, HLA-A, ITGB2, ICAM1, ITGB1, HLA-C, CDH1, IFITM1, HLA-E, HLA-G, CD34, CD81, B2M, CXADR
#> REACTOME_RNA_POLYMERASE_I_PROMOTER_CLEARANCE                                                                                                      POLR1C, POLR1A, POLR1D, UBTF, MAPK3, CCNH, MNAT1, POLR2H
#> REACTOME_G_ALPHA_I_SIGNALLING_EVENTS                                                                               GNG7, AGT, C3, C5, NPY, ANXA1, GNAI2, GNAI3, CXCL12, GNG2, SST, GNB1, GNB2, GNAI1, GNG5
#> REACTOME_STEROID_METABOLISM                                                       SLC27A2, ABCC3, CYP7B1, TM7SF2, ALB, SCP2, FDFT1, LSS, FABP6, HSD17B4, MVD, HSD17B7, MVK, LBR, EBP, PMVK, NSDHL, CYP51A1
#> REACTOME_G_ALPHA_Q_SIGNALLING_EVENTS                                                      GNG7, UTS2, AGT, GCG, ANXA1, CHRM3, DGKA, PIK3R1, GNA11, GNAQ, GNG2, GNB1, GNB2, GNG5, PLCB3, DGKZ, ITPR2, ITPR1
#> REACTOME_NRAGE_SIGNALS_DEATH_THROUGH_JNK                                                                                                MCF2L, MAPK8, VAV2, RAC1, AKAP13, ABR, ARHGEF7, ARHGEF16, ARHGEF18
#>                                                                                   ingroup_mean
#> REACTOME_IMMUNOREGULATORY_INTERACTIONS_BETWEEN_A_LYMPHOID_AND_A_NON_LYMPHOID_CELL    0.6957938
#> REACTOME_RNA_POLYMERASE_I_PROMOTER_CLEARANCE                                        -0.7181860
#> REACTOME_G_ALPHA_I_SIGNALLING_EVENTS                                                 0.4795099
#> REACTOME_STEROID_METABOLISM                                                          0.7851146
#> REACTOME_G_ALPHA_Q_SIGNALLING_EVENTS                                                 0.4246414
#> REACTOME_NRAGE_SIGNALS_DEATH_THROUGH_JNK                                            -0.7748090
#>                                                                                   outgroup_n
#> REACTOME_IMMUNOREGULATORY_INTERACTIONS_BETWEEN_A_LYMPHOID_AND_A_NON_LYMPHOID_CELL       2986
#> REACTOME_RNA_POLYMERASE_I_PROMOTER_CLEARANCE                                            2986
#> REACTOME_G_ALPHA_I_SIGNALLING_EVENTS                                                    2986
#> REACTOME_STEROID_METABOLISM                                                             2986
#> REACTOME_G_ALPHA_Q_SIGNALLING_EVENTS                                                    2986
#> REACTOME_NRAGE_SIGNALS_DEATH_THROUGH_JNK                                                2986
#>                                                                                   outgroup_mean
#> REACTOME_IMMUNOREGULATORY_INTERACTIONS_BETWEEN_A_LYMPHOID_AND_A_NON_LYMPHOID_CELL  -0.003748384
#> REACTOME_RNA_POLYMERASE_I_PROMOTER_CLEARANCE                                        0.001929311
#> REACTOME_G_ALPHA_I_SIGNALLING_EVENTS                                               -0.002420952
#> REACTOME_STEROID_METABOLISM                                                        -0.004761477
#> REACTOME_G_ALPHA_Q_SIGNALLING_EVENTS                                               -0.002575319
#> REACTOME_NRAGE_SIGNALS_DEATH_THROUGH_JNK                                            0.002342385
#>                                                                                       zscore
#> REACTOME_IMMUNOREGULATORY_INTERACTIONS_BETWEEN_A_LYMPHOID_AND_A_NON_LYMPHOID_CELL  0.9340115
#> REACTOME_RNA_POLYMERASE_I_PROMOTER_CLEARANCE                                      -0.5846516
#> REACTOME_G_ALPHA_I_SIGNALLING_EVENTS                                               0.5242025
#> REACTOME_STEROID_METABOLISM                                                        0.8686835
#> REACTOME_G_ALPHA_Q_SIGNALLING_EVENTS                                               0.4936846
#> REACTOME_NRAGE_SIGNALS_DEATH_THROUGH_JNK                                          -0.7970220
#>                                                                                   oddsratio
#> REACTOME_IMMUNOREGULATORY_INTERACTIONS_BETWEEN_A_LYMPHOID_AND_A_NON_LYMPHOID_CELL  7.163636
#> REACTOME_RNA_POLYMERASE_I_PROMOTER_CLEARANCE                                      -6.503451
#> REACTOME_G_ALPHA_I_SIGNALLING_EVENTS                                               6.379465
#> REACTOME_STEROID_METABOLISM                                                        6.181998
#> REACTOME_G_ALPHA_Q_SIGNALLING_EVENTS                                               5.320742
#> REACTOME_NRAGE_SIGNALS_DEATH_THROUGH_JNK                                          -5.311823
#>                                                                                         pvalue
#> REACTOME_IMMUNOREGULATORY_INTERACTIONS_BETWEEN_A_LYMPHOID_AND_A_NON_LYMPHOID_CELL 7.856505e-13
#> REACTOME_RNA_POLYMERASE_I_PROMOTER_CLEARANCE                                      7.849792e-11
#> REACTOME_G_ALPHA_I_SIGNALLING_EVENTS                                              1.777083e-10
#> REACTOME_STEROID_METABOLISM                                                       6.329517e-10
#> REACTOME_G_ALPHA_Q_SIGNALLING_EVENTS                                              1.033449e-07
#> REACTOME_NRAGE_SIGNALS_DEATH_THROUGH_JNK                                          1.085339e-07
#>                                                                                      BH_pvalue
#> REACTOME_IMMUNOREGULATORY_INTERACTIONS_BETWEEN_A_LYMPHOID_AND_A_NON_LYMPHOID_CELL 4.886746e-10
#> REACTOME_RNA_POLYMERASE_I_PROMOTER_CLEARANCE                                      2.441285e-08
#> REACTOME_G_ALPHA_I_SIGNALLING_EVENTS                                              3.684485e-08
#> REACTOME_STEROID_METABOLISM                                                       9.842398e-08
#> REACTOME_G_ALPHA_Q_SIGNALLING_EVENTS                                              1.125135e-05
#> REACTOME_NRAGE_SIGNALS_DEATH_THROUGH_JNK                                          1.125135e-05
#>                                                                                   SignedBH_pvalue
#> REACTOME_IMMUNOREGULATORY_INTERACTIONS_BETWEEN_A_LYMPHOID_AND_A_NON_LYMPHOID_CELL    4.886746e-10
#> REACTOME_RNA_POLYMERASE_I_PROMOTER_CLEARANCE                                        -2.441285e-08
#> REACTOME_G_ALPHA_I_SIGNALLING_EVENTS                                                 3.684485e-08
#> REACTOME_STEROID_METABOLISM                                                          9.842398e-08
#> REACTOME_G_ALPHA_Q_SIGNALLING_EVENTS                                                 1.125135e-05
#> REACTOME_NRAGE_SIGNALS_DEATH_THROUGH_JNK                                            -1.125135e-05
#>                                                                                   background_n
#> REACTOME_IMMUNOREGULATORY_INTERACTIONS_BETWEEN_A_LYMPHOID_AND_A_NON_LYMPHOID_CELL           NA
#> REACTOME_RNA_POLYMERASE_I_PROMOTER_CLEARANCE                                                NA
#> REACTOME_G_ALPHA_I_SIGNALLING_EVENTS                                                        NA
#> REACTOME_STEROID_METABOLISM                                                                 NA
#> REACTOME_G_ALPHA_Q_SIGNALLING_EVENTS                                                        NA
#> REACTOME_NRAGE_SIGNALS_DEATH_THROUGH_JNK                                                    NA
#>                                                                                   background_mean
#> REACTOME_IMMUNOREGULATORY_INTERACTIONS_BETWEEN_A_LYMPHOID_AND_A_NON_LYMPHOID_CELL              NA
#> REACTOME_RNA_POLYMERASE_I_PROMOTER_CLEARANCE                                                   NA
#> REACTOME_G_ALPHA_I_SIGNALLING_EVENTS                                                           NA
#> REACTOME_STEROID_METABOLISM                                                                    NA
#> REACTOME_G_ALPHA_Q_SIGNALLING_EVENTS                                                           NA
#> REACTOME_NRAGE_SIGNALS_DEATH_THROUGH_JNK                                                       NA