Open the compiled spatial model in SLiM

When run, the compiled SLiM script will save the location of each individual that ever lived, and will also record a sample of ancient individuals from each population in a tree sequence data structure which will be saved for all "present-day" individuals at the end of the simulation. This obviously does not make sense for all potential uses and the exact specification of output formats will be changed at some point soon.

run(
  model_dir,
  gen_time,
  burnin,
  sim_length,
  seq_length,
  recomb_rate,
  max_distance,
  max_spread,
  track_ancestry = FALSE,
  output_prefix = file.path(normalizePath(model_dir), "output_"),
  ...,
  include = NULL
)

Arguments

model_dir
Directory where compile saved all spatial maps and other model configuration files

gen_time
Generation time (in model's time units, i.e. years)

burnin
Length of the burnin (in model's time units, i.e. years)

sim_length
Total length of the simulation (in model's time units, i.e. years)

seq_length
Total length of the simulated sequence in base-pairs

recomb_rate
Recombination rate of the simulated sequence

max_distance, max_spread
Maximum values for spatial interaction and mating choice (max_distance) and maximum distance for offspring from its parents (max_spread)

track_ancestry
Track ancestry proportion dynamics in all populations throughout the simulations (default FALSE)? If a non-zero integer is provided, ancestry will be tracked using the number number of neutral ancestry markers equal to this number.

output_prefix
Directory and shared prefix of all output files (all output files will be placed into the model directory by default)